requirement.txt
altgraph==0.16.1
appnope==0.1.0
asn1crypto==1.0.1
astroid==2.3.1
attrs==19.3.0
azure-common==1.1.23
azure-nspkg==3.0.2
azure-storage==0.36.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.0
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.7
django-cors-headers==3.4.0
djangorestframework==3.8.2
djongo==1.3.3
entrypoints==0.3
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
-e git+https://github.com/gauravmm/jupyter-testing.git@01dd39fa2e67b3bf685d623fd3d2cf71a85b2c0c#egg=jupyter_testing
kazoo==2.5.0
Lasagne==0.1
lazy-object-proxy==1.4.2
macholib==1.11
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
modulegraph==0.17
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py2app==0.19
pycparser==2.19
Pygments==2.4.2
pykafka==2.8.0
pylint==2.4.2
pymongo==3.11.0
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2020.1
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
soupsieve==2.0.1
sqlparse==0.2.4
tabulate==0.8.7
terminado==0.8.2
testpath==0.4.4
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
